added separate AppConfiguration class for MessageHelper, removed @ResponseBody for TesAdapterController class
