+++ tes-adapter/src/main/java/com/epam/pipeline/tesadapter/configuration/AppConfiguration.java
@@ -1,0 +1,23 @@
+package com.epam.pipeline.tesadapter.configuration;
+
+import com.epam.pipeline.tesadapter.common.MessageHelper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ResourceBundleMessageSource;
+
+@Configuration
+public class AppConfiguration {
+
+    @Bean
+    public MessageHelper messageHelper() {
+        return new MessageHelper(messageSource());
+    }
+
+    @Bean
+    public ResourceBundleMessageSource messageSource() {
+        ResourceBundleMessageSource messageSource = new ResourceBundleMessageSource();
+        messageSource.setBasename("messages");
+        messageSource.setDefaultEncoding("UTF-8");
+        return messageSource;
+    }
+}
