+++ tes-adapter/src/main/java/com/epam/pipeline/tesadapter/controller/TesAdapterController.java
@@ -14,7 +14,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -43,13 +42,11 @@ public class TesAdapterController {
     }
 
     @PostMapping("/v1/tasks")
-    @ResponseBody
     ResponseEntity<TesCreateTaskResponse> submitTesTask(@RequestBody TesTask body) {
         return ResponseEntity.status(HttpStatus.OK).body(tesTaskService.submitTesTask(body));
     }
 
     @GetMapping("/v1/tasks/{id}")
-    @ResponseBody
     ResponseEntity<TesTask> getTesTask(@PathVariable String id, @RequestParam(required = false,
             defaultValue = "MINIMAL") TaskView view) {
         tesTaskService.stub();
@@ -57,7 +54,6 @@ public class TesAdapterController {
     }
 
     @PostMapping("/v1/tasks/{id}:cancel")
-    @ResponseBody
     ResponseEntity<TesCancelTaskResponse> cancelTesTask(@PathVariable String id) {
         return ResponseEntity.status(HttpStatus.OK).body(tesTaskService.cancelTesTask(id));
     }
